+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.1.1
 
+- Refactor adding wspay payment.
 - Add support for djangoSHOP 0.11.x.
 
 ## 0.1.0